add: mongo proxy core-shell
